Added clothings and fixed male legs
Re-added demon clothing that got deleted

Guys I'm gonna give you all a crash course on source contol if you still use it like this in future projects because this is absolute cancer. We even had a lecture on this from Aaron the year 1 programming teacher... please.

You all had the opportunity to learn it the hard way quickly throughout this project, but you all decided to learn it slowly, and not even completely at that. I've had enough of getting my stuff deleted or overwritten because certain people didn't listen to a lecture and/or bother using separate branches for themselves way earlier. Zai jien